vcaa_project/backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
PyMySQL==1.1.0
python-dotenv==1.0.1
openai==1.12.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
